chore(queue): move execution log to its own package
